requirements.txt
requests==2.31.0
aiohttp==3.9.1
apscheduler==3.10.4
python-dotenv==1.0.0
loguru==0.7.2
PyYAML==6.0.1